refactor(auto_loop): 重构自动循环模块的架构设计
- 将分散的 prompt 模板统一管理到新的 AutoLoopPromptComposer 类
- 优化 AutoLoopEngine 的公共接口设计，添加只读属性和状态写入方法
- 在 AutoLoopWorker 中集成新的 prompt 组合器并更新调用逻辑
- 统一步骤完成检测和文档更新检查逻辑到 Engine 层
- 优化 Worker 中的状态管理，使用新的属性访问方式
- 清理重复的正则表达式导入和阶段约束常量定义
- 添加组件间调用的封装，提高代码模块化程度
